Move GlobeBrowsing thread pool to OpenSpace core
